checking the resource that triggered the event (police_impound)

+++ server/police_impound.lua
@@ -42,12 +42,14 @@ end)
 
 --- events
 RegisterNetEvent('rhd_garage:server:removeFromPoliceImpound', function( plate )
+    if GetInvokingResource() then return end
     MySQL.query('DELETE FROM police_impound WHERE plate = ?', { plate })
     MySQL.update(DBFormat.getParameters("state_policeImpound_remove"), DBFormat.getValue("state_policeImpound_remove", plate))
 end)
 
 RegisterNetEvent('rhd_garage:server:policeImpound.sendBill', function( citizenid, fine, plate )
     local src
+    if GetInvokingResource() then return end
     local Player = Framework.server.GetPlayerFromCitizenid(citizenid)
     if not Player then return end
     if Framework.esx() then src = Player.source elseif Framework.qb() then src = Player.PlayerData.source end
